+++ modules/DataContractDsl/org.mod4j.dsl.datacontract.generator/src/main/java/org/mod4j/dsl/datacontract/generator/helpers/DtoHelpers.java
@@ -81,7 +81,7 @@ public class DtoHelpers {
 
     public static String javaType(DtoDecimalProperty p) {
         // return p.isNullable() ? "Float" : "float";
-        return "Float";
+        return "Double";
     }
 
     public static List<DtoProperty> getAllProperties(BusinessClassDto cls) {
